+++ Material.hpp
@@ -6,7 +6,7 @@
 class Material
 {
 public:
-    virtual constexpr float GetAlbedo(Coord2D uv) const = 0;
-    virtual constexpr float GetOpacity(Coord2D uv) const = 0;
+    virtual constexpr double GetAlbedo(Coord2D uv) const = 0;
+    virtual constexpr double GetOpacity(Coord2D uv) const = 0;
     virtual constexpr Color GetColor(Coord2D uv) const = 0;
 };
