+++ Maths.ipp
@@ -1,13 +1,13 @@
 namespace Maths
 {
-constexpr float abs(float number)
+constexpr double abs(double number)
 {
-    return number < 0.f ? -number : number;
+    return number < 0.0 ? -number : number;
 }
 
-constexpr float pow(float number, int n)
+constexpr double pow(double number, int n)
 {
-    float result = 1.f;
+    double result = 1.0;
 
     while (n > 0)
     {
@@ -32,31 +32,31 @@ constexpr int fact(int n)
 
 // Newton-Raphson method.
 // https://en.wikipedia.org/wiki/Newton's_method
-constexpr float sqrt(float n)
+constexpr double sqrt(double n)
 {
-    float result = n;
-    float previous = 0.f;
+    double result = n;
+    double previous = 0.0;
     while (result != previous)
     {
         previous = result;
-        result = (result + n / result) / 2.f;
+        result = (result + n / result) / 2.0;
     }
     return result;
 }
 
-constexpr float degToRad(float angle)
+constexpr double degToRad(double angle)
 {
-    return angle * _PI / 180.f;
+    return angle * _PI / 180.0;
 }
 
-constexpr float cos(float angle)
+constexpr double cos(double angle)
 {
     return (sin(angle + _PI_2));
 }
 
 // Taylor's polynomial theorem with 5 iterations.
 // https://en.wikipedia.org/wiki/Taylor_series#Approximation_error_and_convergence
-constexpr float sin(float angle)
+constexpr double sin(double angle)
 {
     // set x between _PI and -_PI.
     while (angle < -_PI)
@@ -75,7 +75,7 @@ constexpr float sin(float angle)
     // clang-format on
 }
 
-constexpr float acos(float number)
+constexpr double acos(double number)
 {
     return (_PI_2 - asin(number));
 }
@@ -83,15 +83,15 @@ constexpr float acos(float number)
 // Will only work with number between -1 and 1 !
 // Here we use it only for UV functions.
 // Quite imprecise for number close of -1 or 1.
-constexpr float asin(float number)
+constexpr double asin(double number)
 {
     // clang-format off
     return (number
-            + pow(number, 3) * 1.f / 6.f
-            + pow(number, 5) * 3.f / 40.f
-            + pow(number, 7) * 5.f / 112.f
-            + pow(number, 9) * 35.f / 1152.f
-            + pow(number, 11) * 63.f / 2816.f
+            + pow(number, 3) * 1.0 / 6.0
+            + pow(number, 5) * 3.0 / 40.0
+            + pow(number, 7) * 5.0 / 112.0
+            + pow(number, 9) * 35.0 / 1152.0
+            + pow(number, 11) * 63.0 / 2816.0
             );
     // clang-format on
 }
