+++ Quadratic.hpp
@@ -6,11 +6,11 @@
 
 namespace Quadratic
 {
-constexpr int solve(float a, float b, float c, float& result1, float& result2);
-constexpr std::pair<Optional<float>, Optional<float>> sortedPositiveResults(float a, float b, float c);
+constexpr int solve(double a, double b, double c, double& result1, double& result2);
+constexpr std::pair<Optional<double>, Optional<double>> sortedPositiveResults(double a, double b, double c);
 } // namespace Quadratic
 
 #include "Quadratic.ipp"
 
-static_assert(*Quadratic::sortedPositiveResults(3, -4, 1).first == 1.f / 3.f);
-static_assert(*Quadratic::sortedPositiveResults(-2, 3, 2).first == 2.f);
+static_assert(*Quadratic::sortedPositiveResults(3, -4, 1).first == 1.0 / 3.0);
+static_assert(*Quadratic::sortedPositiveResults(-2, 3, 2).first == 2.0);
