+++ cameras/FishEyeCamera.hpp
@@ -5,14 +5,14 @@
 class FishEyeCamera final : public Camera
 {
 public:
-    constexpr FishEyeCamera(const Vector& position, const Matrix& rotation, float azimuthAngle, float siteAngle);
+    constexpr FishEyeCamera(const Vector& position, const Matrix& rotation, double azimuthAngle, double siteAngle);
 
     // Both of those params are between -0.5 and 0.5
-    constexpr Ray GetRay(float pixelX, float pixelY) const override;
+    constexpr Ray GetRay(double pixelX, double pixelY) const override;
 
 private:
-    float m_azimuthAngle;
-    float m_siteAngle;
+    double m_azimuthAngle;
+    double m_siteAngle;
 };
 
 #include "FishEyeCamera.ipp"
